+++ lib/src/models/block_override.dart
@@ -1,5 +1,6 @@
 import 'package:block_dialog/src/animations/block_animation/block_animation.dart';
 import 'package:block_dialog/src/layout/blocks/block.dart';
+import 'package:flutter/material.dart';
 
 class BlockOverride {
   /// Override the block position used for border radius.
@@ -10,6 +11,8 @@ class BlockOverride {
 
   /// Use a custom animation for this block.
   BlockAnimation? customAnimation;
+
+  EdgeInsetsGeometry? padding;
 
   /// Construct a block override.
   BlockOverride({